Add files via upload
This is the c program to insert and delete the node in singly linked list, In this program only once we can create a link list and if we perform the delete or insert operation we cannot add any node using the create list. Hence, we have to either use insert or delete the nodes operation to manipulate the list.  When i get time again i will definately rectify the issue. But for now this program is also working fine.